shift model stub
