+++ src/test/java/seedu/address/logic/commands/AddTuteeCommandTest.java
@@ -4,30 +4,24 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
-import java.util.function.Predicate;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import javafx.collections.ObservableList;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.UndoRedoStack;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.Task;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.person.exceptions.PersonNotFoundException;
-import seedu.address.model.tag.Tag;
 import seedu.address.model.tutee.Tutee;
+import seedu.address.testutil.ModelStub;
 import seedu.address.testutil.TuteeBuilder;
 
 //@@author ChoChihTun
@@ -55,7 +49,7 @@ public class AddTuteeCommandTest {
 
     @Test
     public void execute_duplicatePerson_throwsCommandException() throws Exception {
-        ModelStub modelStub = new ModelStubThrowingDuplicatePersonException();
+        ModelStub  = new ModelStubThrowingDuplicatePersonException();
         Tutee validTutee = new TuteeBuilder().build();
 
         thrown.expect(CommandException.class);
@@ -98,88 +92,9 @@ public class AddTuteeCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void addPerson(Person person) throws DuplicatePersonException {
-            fail("This method should not be called.");
-        }
-
-        @Override
-        public void addTask(Task task) {
-            fail("This method should not be called");
-        }
-
-        @Override
-        public void deleteTask(Task task) {
-            fail("This method should not be called");
-        }
-
-        @Override
-        public void updateTask(Task task, Task editedTask) {
-            fail("This method should not be called");
-        }
-
-        @Override
-        public ObservableList<Task> getFilteredTaskList() {
-            fail("This method should not be called.");
-            return null;
-        }
-
-        @Override
-        public void updateFilteredTaskList(Predicate<Task> predicate) {
-            fail("This method should not be called.");
-        }
-
-        @Override
-        public void resetData(ReadOnlyAddressBook newData) {
-            fail("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            fail("This method should not be called.");
-            return null;
-        }
-
-        @Override
-        public void deletePerson(Person target) throws PersonNotFoundException {
-            fail("This method should not be called.");
-        }
-
-        @Override
-        public void updatePerson(Person target, Person editedPerson)
-                throws DuplicatePersonException {
-            fail("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Person> getFilteredPersonList() {
-            fail("This method should not be called.");
-            return null;
-        }
-
-        @Override
-        public void updateFilteredPersonList(Predicate<Person> predicate) {
-            fail("This method should not be called.");
-        }
-
-        @Override
-        public void deleteTag(Tag tag, Person person) {
-            fail("deleteTag should not be called when adding Person.");
-        }
-
-        @Override
-        public void sortFilteredPersonList(Comparator<Person> comparator) {
-            fail("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that always throw a DuplicatePersonException when trying to add a person.
      */
-    private class ModelStubThrowingDuplicatePersonException extends AddTuteeCommandTest.ModelStub {
+    private class ModelStubThrowingDuplicatePersonException extends ModelStub {
         @Override
         public void addPerson(Person person) throws DuplicatePersonException {
             throw new DuplicatePersonException();
@@ -194,7 +109,7 @@ public class AddTuteeCommandTest {
     /**
      * A Model stub that always accept the person being added.
      */
-    private class ModelStubAcceptingPersonAdded extends AddTuteeCommandTest.ModelStub {
+    private class ModelStubAcceptingPersonAdded extends ModelStub {
         final ArrayList<Person> personsAdded = new ArrayList<>();
 
         @Override
